refactor test for pytorch quantization
